# docs/swagger.yaml
definitions:
...
  operation.CreateTeamOP:
    properties:
      team_name:
        example: sleekflow
        maxLength: 15
        minLength: 3
        type: string
    required:
    - team_name
    type: object
...
  operation.UpdateTeamOP:
    properties:
      team_name:
        example: sleekflow
        maxLength: 15
        minLength: 3
        type: string
      user:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
    type: object
  operation.UpdateUserOP:
    properties:
      password:
        example: password
        maxLength: 16
        minLength: 5
        type: string
      user_name:
        example: admin
        maxLength: 15
        minLength: 3
        type: string
    required:
    - password
    - user_name
    type: object
  operation.UserOP:
    properties:
      password:
        example: password
        maxLength: 16
        minLength: 5
        type: string
      user_name:
        example: admin
        maxLength: 15
        minLength: 3
        type: string
    required:
    - password
    - user_name
    type: object
...
  serializer.ResponseTeam:
    properties:
      data:
        $ref: '#/definitions/serializer.Team'
      error:
        type: string
      msg:
        type: string
      status:
        type: integer
    type: object
  serializer.ResponseTeamMembers:
    properties:
      data:
        items:
          $ref: '#/definitions/serializer.User'
        type: array
      error:
        type: string
      msg:
        type: string
      status:
        type: integer
    type: object
  serializer.ResponseUser:
    properties:
      data:
        $ref: '#/definitions/serializer.User'
      error:
        type: string
      msg:
        type: string
      status:
        type: integer
    type: object
  serializer.ResponseUserTeams:
    properties:
      data:
        items:
          $ref: '#/definitions/serializer.Team'
        type: array
      error:
        type: string
      msg:
        type: string
      status:
        type: integer
    type: object
...
  serializer.Team:
    properties:
      create_at:
        type: integer
      id:
        example: 1
        type: integer
      team_name:
        example: Sleekflow
        type: string
    type: object
  serializer.User:
    properties:
      create_at:
        type: integer
      id:
        example: 1
        type: integer
      user_name:
        example: admin
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/team/:team_id:
    delete:
      consumes:
      - application/json
      parameters:
      - description: team id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"status":200,"data":{},"msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTeam'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTeam'
      summary: Delete team
    get:
      consumes:
      - application/json
      parameters:
      - description: team id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"status":200,"data":{},"msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTeamMembers'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTeamMembers'
      summary: Show team members
    put:
      consumes:
      - application/json
      parameters:
      - description: team id
        in: path
        name: id
        required: true
        type: integer
      - description: team info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/operation.UpdateTeamOP'
      produces:
      - application/json
      responses:
        "200":
          description: '{"status":200,"data":{},"msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTeam'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTeam'
      summary: Update team info
...
  /api/v1/user:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":true,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseUserTeams'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseUserTeams'
      summary: Update user info
    put:
      consumes:
      - application/json
      parameters:
      - description: username, password
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/operation.UpdateUserOP'
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":true,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseUserTeams'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseUserTeams'
      summary: Update user info
  /api/v1/user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: user_name, password
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/operation.UserOP'
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":true,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseUser'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseUser'
      summary: User login
  /api/v1/user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: username, password
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/operation.UserOP'
      produces:
      - application/json
      responses:
        "200":
          description: '{"status":200,"data":{},"msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseUser'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseUser'
      summary: User register
...
  /api/v1/user/team:
    post:
      consumes:
      - application/json
      parameters:
      - description: team name
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/operation.CreateTeamOP'
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":true,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTeam'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseTeam'
      summary: Create team
  /api/v1/user/team/:team_id:
    delete:
      consumes:
      - application/json
      parameters:
      - description: team_id
        in: path
        name: team_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":true,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseUserTeams'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseUserTeams'
      summary: Leave team
    put:
      consumes:
      - application/json
      parameters:
      - description: team_id
        in: path
        name: team_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":true,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseUserTeams'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseUserTeams'
      summary: Join team
  /api/v1/user/teams:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"Status":true,"data":{},"Msg":"ok"}'
          schema:
            $ref: '#/definitions/serializer.ResponseUserTeams'
        "500":
          description: '{"status":500,"data":{},"Msg":{},"Error":"error"}'
          schema:
            $ref: '#/definitions/serializer.ResponseUserTeams'
      summary: Show user's teams
...
swagger: "2.0"